Fixed unit tests

+++ test/test-utils/lib/db.js
@@ -6,10 +6,17 @@ var async = require('async');
 var DB = NOOT.Namespace.create({
 
   create: function(options, callback) {
-    var db = mongoose.createConnection();
-    db.open('mongodb://localhost:27017/' + options.name, function() {
-      if (options.drop) return db.db.dropDatabase(callback);
-      else return callback();
+    var hasFirstConnected = false;
+    var db = mongoose.createConnection('mongodb://localhost:27017/' + options.name);
+    db.on('error', function(err) {
+      throw err;
+    });
+    db.on('connected', function() {
+      if (!hasFirstConnected) {
+        hasFirstConnected = true;
+        if (options.drop) return db.db.dropDatabase(callback);
+        else return callback();
+      }
     });
     return db;
   }
